Added Scratch-Off
Akihabara, the electric town--hub for state-of-the-art technology, fashion, otaku goods, and the cutest boys in skirts you will ever see. Pop into the Girls! Girls! Girls!? Cafe and fall in love with one of four charming "ladies..." or maybe even all of them? Jump into a story full of fluff, romance, friendship, and crossdressing that will keep you coming back for more!